# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Comprehensive API
  version: 1.0.0
  description: API documentation for user accounts, departments, employees, requests, and workflows.
servers:
  - url: http://localhost:3000 # Adjust if your port is different
...
paths:
...
  /departments/:
    get:
      tags: [Departments]
      summary: Get all departments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentDetailed'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Departments]
      summary: Create a new department (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreateOrUpdate'
      responses:
        '201':
          description: Department created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Department created successfully.
        '400':
          description: Validation error (e.g., name required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /departments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the department
        schema:
          type: integer
    get:
      tags: [Departments]
      summary: Get department by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Department data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDetailed'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Departments]
      summary: Update department by ID (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreateOrUpdate'
      responses:
        '200':
          description: Department updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Department updated successfully.
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Departments]
      summary: Delete department by ID (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      responses:
        '200':
          description: Department deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Department deleted successfully.
        '400':
          description: Cannot delete department with assigned employees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Employee Endpoints ---
...
  /workflows/:
    post:
      tags: [Workflows]
      summary: Create a generic workflow (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreateGeneric'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow' # Returns the created workflow
        '400':
          description: Validation error or employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows/employee/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the employee
        schema:
          type: integer
    get:
      tags: [Workflows]
      summary: Get all workflows for a specific employee
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of workflows for the employee
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workflow'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (user cannot view these workflows)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the workflow
        schema:
          type: integer
    put:
      tags: [Workflows]
      summary: Update workflow status by ID (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: Approved
              required: [status]
      responses:
        '200':
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Workflow updated successfully.
        '400':
          description: Validation error (status required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Workflows]
      summary: Delete workflow by ID (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      responses:
        '200':
          description: Workflow deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Workflow deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows/onboarding:
    post:
      tags: [Workflows]
      summary: Create an onboarding workflow (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: integer
              required: [employeeId]
      responses:
        '201':
          description: Onboarding workflow created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Workflow onboarding created successfully.
        '400':
          description: Employee ID required or employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows/transfer:
    post:
      tags: [Workflows]
      summary: Create a department transfer workflow (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: integer
                oldDepartment:
                  type: object
                  properties:
                    name:
                      type: string
                newDepartment:
                  type: object
                  properties:
                    name:
                      type: string
              required: [employeeId, oldDepartment, newDepartment]
      responses:
        '201':
          description: Transfer workflow created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Workflow transfer created successfully.
        '400':
          description: Required fields missing or employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows/leave:
    post:
      tags: [Workflows]
      summary: Create a leave request workflow (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: integer
                startDate:
                  type: string
                  format: date
                  example: "2024-12-20"
                endDate:
                  type: string
                  format: date
                  example: "2024-12-24"
              required: [employeeId, startDate, endDate]
      responses:
        '201':
          description: Leave workflow created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Workflow request leave created successfully.
        '400':
          description: Required fields missing or employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows/resources:
    post:
      tags: [Workflows]
      summary: Create a resources request workflow (Admin only)
      security:
        - bearerAuth: [] # Role.Admin enforced by fake backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: integer
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/RequestItemBase'
              required: [employeeId, items]
      responses:
        '201':
          description: Resources workflow created successfully
          content:
            application/json:
              schema: # Fake backend returns message, node service returns message + workflowId
                type: object
                properties:
                  message:
                    type: string
                    example: Workflow request item created successfully. # Matched to Node.js controller for /resources
                  # workflowId: # If matching Node.js service more closely
                  #   type: integer
        '400':
          description: Required fields missing, invalid items, or employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    DepartmentBase:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        description:
          type: string
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          nullable: true
          readOnly: true
      required: [name, description]
    DepartmentDetailed:
      allOf:
        - $ref: '#/components/schemas/DepartmentBase'
        - type: object
          properties:
            employeeCount:
              type: integer
              readOnly: true
    DepartmentCreateOrUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [name]


    # --- Employee Schemas ---
...
    RequestItemBase:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: integer
          minimum: 1
      required: [name, quantity]
...
    Workflow:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        employeeId:
          type: integer
        type:
          type: string
          example: Onboarding # Or "Department Transfer", "Request Approval", "Resources Request" etc.
        details:
          type: string
        status:
          type: string
          enum: [Pending, InProgress, Completed, Failed, Cancelled]
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          nullable: true
          readOnly: true
      required: [employeeId, type, details, status]
    WorkflowCreateGeneric: # For POST /workflows/
      type: object
      properties:
        employeeId:
          type: integer
        type:
          type: string
        details:
          type: string
        status: # Optional, defaults to Pending
          type: string
          enum: [Pending, InProgress, Completed, Failed, Cancelled]
          nullable: true
      required: [employeeId, type, details]
    Error:
      type: object
      properties:
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT